backend/requirements.txt
asgiref==3.3.4
async-timeout==3.0.1
Django==3.2.3
django-admin-list-filter-dropdown==1.0.3
django-filter==2.4.0
djangorestframework==3.12.4
Jinja2==2.10.1
Markdown==3.3.4
MarkupSafe==1.1.0
PyMySQL==1.0.2
PyNaCl==1.3.0
requests==2.22.0
requests-unixsocket==0.2.0
simplejson==3.16.0
python-dotenv==0.17.0
